+++ client/control.go
@@ -2,12 +2,12 @@ package client
 
 import (
 	"bufio"
+	"context"
 	"crypto/tls"
 	"encoding/base64"
 	"encoding/binary"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"math"
 	"math/rand"
@@ -147,7 +147,7 @@ re:
 		}
 		vkey = string(b)
 	}
-	ioutil.WriteFile(filepath.Join(common.GetTmpPath(), "npc_vkey.txt"), []byte(vkey), 0600)
+	os.WriteFile(filepath.Join(common.GetTmpPath(), "npc_vkey.txt"), []byte(vkey), 0600)
 
 	//send hosts to server
 	for _, v := range cnf.Hosts {
@@ -343,6 +343,7 @@ func handleP2PUdp(localAddr, rAddr, md5Password, role string) (remoteAddress str
 	if err != nil {
 		return
 	}
+	defer localConn.Close()
 	err = getRemoteAddressFromServer(rAddr, localConn, md5Password, role, 0)
 	if err != nil {
 		logs.Error(err)
@@ -359,6 +360,7 @@ func handleP2PUdp(localAddr, rAddr, md5Password, role string) (remoteAddress str
 		return
 	}
 	var remoteAddr1, remoteAddr2, remoteAddr3 string
+	localConn.SetReadDeadline(time.Now().Add(time.Second * 30))
 	for {
 		buf := make([]byte, 1024)
 		if n, addr, er := localConn.ReadFromUDP(buf); er != nil {
@@ -390,8 +392,8 @@ func handleP2PUdp(localAddr, rAddr, md5Password, role string) (remoteAddress str
 func sendP2PTestMsg(localConn *net.UDPConn, remoteAddr1, remoteAddr2, remoteAddr3 string) (string, error) {
 	logs.Trace(remoteAddr3, remoteAddr2, remoteAddr1)
 	defer localConn.Close()
-	isClose := false
-	defer func() { isClose = true }()
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
 	interval, err := getAddrInterval(remoteAddr1, remoteAddr2, remoteAddr3)
 	if err != nil {
 		return "", err
@@ -411,8 +413,10 @@ func sendP2PTestMsg(localConn *net.UDPConn, remoteAddr1, remoteAddr2, remoteAddr
 		for {
 			select {
 			case <-ticker.C:
-				if isClose {
+				select {
+				case <-ctx.Done():
 					return
+				default:
 				}
 				if _, err := localConn.WriteTo([]byte(common.WORK_P2P_CONNECT), remoteUdpAddr); err != nil {
 					return
@@ -437,8 +441,10 @@ func sendP2PTestMsg(localConn *net.UDPConn, remoteAddr1, remoteAddr2, remoteAddr
 					for {
 						select {
 						case <-ticker.C:
-							if isClose {
+							select {
+							case <-ctx.Done():
 								return
+							default:
 							}
 							if _, err := localConn.WriteTo([]byte(common.WORK_P2P_CONNECT), remoteUdpAddr); err != nil {
 								return
@@ -552,13 +558,13 @@ func getRandomPortArr(min, max int) []int {
 	for i := min; i <= max; i++ {
 		addrAddr[max-i] = i
 	}
-	rand.Seed(time.Now().UnixNano())
-	var r, temp int
+	r := rand.New(rand.NewSource(time.Now().UnixNano()))
+	var ri, temp int
 	for i := max - min; i > 0; i-- {
-		r = rand.Int() % i
+		ri = r.Int() % i
 		temp = addrAddr[i]
-		addrAddr[i] = addrAddr[r]
-		addrAddr[r] = temp
+		addrAddr[i] = addrAddr[ri]
+		addrAddr[ri] = temp
 	}
 	return addrAddr
 }