+++ lib/pmux/plistener.go
@@ -10,12 +10,14 @@ type PortListener struct {
 	connCh  chan *PortConn
 	addr    net.Addr
 	isClose bool
+	done    chan struct{}
 }
 
 func NewPortListener(connCh chan *PortConn, addr net.Addr) *PortListener {
 	return &PortListener{
 		connCh: connCh,
 		addr:   addr,
+		done:   make(chan struct{}),
 	}
 }
 
@@ -23,11 +25,15 @@ func (pListener *PortListener) Accept() (net.Conn, error) {
 	if pListener.isClose {
 		return nil, errors.New("the listener has closed")
 	}
-	conn := <-pListener.connCh
-	if conn != nil {
-		return conn, nil
+	select {
+	case <-pListener.done:
+		return nil, errors.New("the listener has closed")
+	case conn := <-pListener.connCh:
+		if conn != nil {
+			return conn, nil
+		}
+		return nil, errors.New("the listener has closed")
 	}
-	return nil, errors.New("the listener has closed")
 }
 
 func (pListener *PortListener) Close() error {
@@ -36,6 +42,7 @@ func (pListener *PortListener) Close() error {
 		return errors.New("the listener has closed")
 	}
 	pListener.isClose = true
+	close(pListener.done)
 	return nil
 }