+++ server/proxy/https.go
@@ -267,7 +267,7 @@ func (https *HttpsServer) handleHttps2(c net.Conn, hostName string, rb []byte, r
 		return
 	}
 	logs.Info("new https connection,clientId %d,host %s,remote address %s", host.Client.Id, r.Host, c.RemoteAddr().String())
-	https.DealClient(conn.NewConn(c), host.Client, targetAddr, rb, common.CONN_TCP, nil, host.Client.Flow, host.Target.LocalProxy, nil)
+	https.DealClient(conn.NewConn(c), host.Client, targetAddr, rb, common.CONN_TCP, nil, host.Client.Flow, host.Target.LocalProxy, nil, host)
 }
 
 // close
@@ -312,7 +312,7 @@ func (https *HttpsServer) handleHttps(c net.Conn) {
 		return
 	}
 	logs.Trace("new https connection,clientId %d,host %s,remote address %s", host.Client.Id, r.Host, c.RemoteAddr().String())
-	https.DealClient(conn.NewConn(c), host.Client, targetAddr, rb, common.CONN_TCP, nil, host.Client.Flow, host.Target.LocalProxy, nil)
+	https.DealClient(conn.NewConn(c), host.Client, targetAddr, rb, common.CONN_TCP, nil, host.Client.Flow, host.Target.LocalProxy, nil, host)
 }
 
 type HttpsListener struct {
@@ -360,7 +360,7 @@ func GetServerNameFromClientHello(c net.Conn) (string, []byte) {
 		return "", header
 	}
 	recordLen := int(header[3])<<8 | int(header[4])
-	if recordLen <= 0 {
+	if recordLen <= 0 || recordLen > 16384 {
 		return "", header
 	}